refactor: create a directory for platform
