+++ packages/rest-api-client/src/platform/node.ts
@@ -1,0 +1,10 @@
+import fs from "fs";
+import { promisify } from "util";
+import { basename } from "path";
+
+const readFile = promisify(fs.readFile);
+export const readFileFromPath = async (filePath: string) => {
+  const data = await readFile(filePath);
+  const name = basename(filePath);
+  return { data, name };
+};
