+++ packages/rest-api-client/src/platform/platformDependencies.ts
